Moved navigationBar to it's own folder
