+++ src/components/navigationBar/navigationBar.scss
@@ -1,4 +1,4 @@
-@import './shared/styles/common.scss';
+@import '../shared/styles/common.scss';
 
 .navigationBarView {
     background-color: $blue;
